Moved Wallet Creation options to the beginning of the Add wallet flow
(cherry picked from commit ab89bc98b4033aaeb12f043818e6893ab6fa616e)
(cherry picked from commit c694abd2a251abdc61f95483eb4d452b3af2bcae)
