+++ WalletWasabi.Fluent/Models/WalletCreationOption.cs
@@ -1,0 +1,9 @@
+﻿namespace WalletWasabi.Fluent.Models;
+
+public enum WalletCreationOption
+{
+	AddNewWallet = 1,
+	ConnectToHardwareWallet = 2,
+	ImportWallet = 3,
+	RecoverWallet = 4
+}
