+++ WalletWasabi.Fluent/ViewModels/AddWallet/AddWalletPageViewModel.cs
@@ -1,13 +1,14 @@
-using ReactiveUI;
 using System.IO;
 using System.Reactive;
 using System.Reactive.Disposables;
-using System.Reactive.Linq;
-using WalletWasabi.Blockchain.Keys;
-using WalletWasabi.Fluent.Validation;
+using System.Threading.Tasks;
+using System.Windows.Input;
+using ReactiveUI;
+using WalletWasabi.Fluent.Helpers;
+using WalletWasabi.Fluent.Models;
 using WalletWasabi.Fluent.ViewModels.Dialogs.Base;
-using WalletWasabi.Models;
-using WalletWasabi.Fluent.ViewModels.NavBar;
+using WalletWasabi.Helpers;
+using WalletWasabi.Logging;
 
 namespace WalletWasabi.Fluent.ViewModels.AddWallet;
 
@@ -22,68 +23,65 @@ namespace WalletWasabi.Fluent.ViewModels.AddWallet;
 	NavBarPosition = NavBarPosition.Bottom)]
 public partial class AddWalletPageViewModel : DialogViewModelBase<Unit>
 {
-	[AutoNotify] private string _walletName = "";
-
 	public AddWalletPageViewModel()
 	{
-		SelectionMode = NavBarItemSelectionMode.Button;
+		CreateWalletCommand = ReactiveCommand.Create(OnCreateWallet);
 
-		var canExecute = this.WhenAnyValue(x => x.WalletName)
-			.ObserveOn(RxApp.MainThreadScheduler)
-			.Select(x => !string.IsNullOrWhiteSpace(x) && !Validations.Any);
+		ConnectHardwareWalletCommand = ReactiveCommand.Create(OnConnectHardwareWallet);
 
-		NextCommand = ReactiveCommand.Create(OnNext, canExecute);
+		ImportWalletCommand = ReactiveCommand.CreateFromTask(async () => await OnImportWalletAsync());
 
-		OpenCommand = ReactiveCommand.Create(
-			async () =>
-		{
-			MainViewModel.Instance.IsOobeBackgroundVisible = true;
+		RecoverWalletCommand = ReactiveCommand.Create(OnRecoverWallet);
 
-			await NavigateDialogAsync(this, NavigationTarget.DialogScreen);
-
-			MainViewModel.Instance.IsOobeBackgroundVisible = false;
-		});
-
-		this.ValidateProperty(x => x.WalletName, errors => ValidateWalletName(errors, WalletName));
+		OpenCommand = ReactiveCommand.Create(async () => await NavigateDialogAsync(this, NavigationTarget.DialogScreen));
 	}
 
-	private void OnNext()
+	public ICommand CreateWalletCommand { get; }
+
+	public ICommand ConnectHardwareWalletCommand { get; }
+
+	public ICommand ImportWalletCommand { get; }
+
+	public ICommand RecoverWalletCommand { get; }
+
+	private void OnCreateWallet()
 	{
-		Navigate().To(new SelectWalletCreationOptionViewModel(WalletName));
+		Navigate().To(new WalletNamePageViewModel(WalletCreationOption.AddNewWallet));
 	}
 
-	private static void ValidateWalletName(IValidationErrors errors, string walletName)
+	private void OnConnectHardwareWallet()
 	{
-		string walletFilePath = Path.Combine(Services.WalletManager.WalletDirectories.WalletsDir, $"{walletName}.json");
+		Navigate().To(new WalletNamePageViewModel(WalletCreationOption.ConnectToHardwareWallet));
+	}
 
-		if (string.IsNullOrEmpty(walletName))
+	private async Task OnImportWalletAsync()
+	{
+		try
 		{
-			return;
+			var filePath = await FileDialogHelper.ShowOpenFileDialogAsync("Import wallet file", new[] { "json" });
+
+			if (filePath is null)
+			{
+				return;
+			}
+
+			var walletName = Path.GetFileNameWithoutExtension(filePath);
+
+			var keyManager = await ImportWalletHelper.ImportWalletAsync(Services.WalletManager, walletName, filePath);
+
+			Navigate().To(new AddedWalletPageViewModel(keyManager));
 		}
-
-		if (walletName.IsTrimmable())
+		catch (Exception ex)
 		{
-			errors.Add(ErrorSeverity.Error, "Leading and trailing white spaces are not allowed!");
-			return;
-		}
-
-		if (File.Exists(walletFilePath))
-		{
-			errors.Add(
-				ErrorSeverity.Error,
-				$"A wallet named {walletName} already exists. Please try a different name.");
-			return;
-		}
-
-		if (!WalletGenerator.ValidateWalletName(walletName))
-		{
-			errors.Add(ErrorSeverity.Error, "Selected Wallet is not valid. Please try a different name.");
+			Logger.LogError(ex);
+			await ShowErrorAsync(Title, ex.ToUserFriendlyString(),
+				"The wallet file was not valid or compatible with Wasabi.");
 		}
 	}
 
-	protected override void OnNavigatedFrom(bool isInHistory)
+	private void OnRecoverWallet()
 	{
-		base.OnNavigatedFrom(isInHistory);
+		Navigate().To(new WalletNamePageViewModel(WalletCreationOption.RecoverWallet));
 	}
 
 	protected override void OnNavigatedTo(bool isInHistory, CompositeDisposable disposables)
@@ -92,12 +90,5 @@ public partial class AddWalletPageViewModel : DialogViewModelBase<Unit>
 
 		var enableCancel = Services.WalletManager.HasWallet();
 		SetupCancel(enableCancel: enableCancel, enableCancelOnEscape: enableCancel, enableCancelOnPressed: enableCancel);
-
-		this.RaisePropertyChanged(WalletName);
-
-		if (!isInHistory)
-		{
-			WalletName = "";
-		}
 	}
 }
