--- WalletWasabi.Fluent/Views/AddWallet/SelectWalletCreationOptionView.axaml
@@ -1,37 +1,0 @@
-<UserControl xmlns="https://github.com/avaloniaui"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:vm="using:WalletWasabi.Fluent.ViewModels.AddWallet"
-             xmlns:c="using:WalletWasabi.Fluent.Controls"
-             mc:Ignorable="d" d:DesignWidth="428" d:DesignHeight="371"
-             x:DataType="vm:SelectWalletCreationOptionViewModel"
-             x:CompileBindings="True"
-             x:Class="WalletWasabi.Fluent.Views.AddWallet.SelectWalletCreationOptionView">
-  <c:ContentArea Title="{Binding Title}"
-                 EnableCancel="{Binding EnableCancel}"
-                 EnableBack="{Binding EnableBack}"
-                 IsBusy="{Binding IsBusy}">
-    <WrapPanel Orientation="Horizontal" Margin="0 0 0 20" HorizontalAlignment="Center">
-      <WrapPanel.Styles>
-        <Style Selector="c|TileButton">
-          <Setter Property="Margin" Value="4"/>
-        </Style>
-      </WrapPanel.Styles>
-      <c:TileButton Text="Create a new wallet"
-                    Classes="action"
-                    Icon="{StaticResource add_regular}"
-                    Command="{Binding CreateWalletCommand}"
-                    IsDefault="{Binding IsActive}" />
-      <c:TileButton Text="Connect to hardware wallet"
-                    Icon="{StaticResource calculator_regular}"
-                    Command="{Binding ConnectHardwareWalletCommand}" />
-      <c:TileButton Text="Import a wallet"
-                    Icon="{StaticResource import_regular}"
-                    Command="{Binding ImportWalletCommand}" />
-      <c:TileButton Text="Recover a wallet"
-                    Icon="{StaticResource recover_arrow_right_regular}"
-                    Command="{Binding RecoverWalletCommand}" />
-    </WrapPanel>
-  </c:ContentArea>
-</UserControl>
