+++ WalletWasabi.Fluent/Views/AddWallet/WalletNamePageView.axaml
@@ -1,0 +1,27 @@
+<UserControl xmlns="https://github.com/avaloniaui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             xmlns:vm="using:WalletWasabi.Fluent.ViewModels.AddWallet"
+             xmlns:c="using:WalletWasabi.Fluent.Controls"
+             xmlns:i="using:Avalonia.Xaml.Interactivity"
+             xmlns:behaviors="using:WalletWasabi.Fluent.Behaviors"
+             mc:Ignorable="d" d:DesignWidth="428" d:DesignHeight="371"
+             x:DataType="vm:WalletNamePageViewModel"
+             x:CompileBindings="True"
+             x:Class="WalletWasabi.Fluent.Views.AddWallet.WalletNamePageView">
+  <c:ContentArea Title="{Binding Title}" Caption="Type in your wallet's name"
+                 EnableCancel="{Binding EnableCancel}"
+                 EnableBack="{Binding EnableBack}"
+                 IsBusy="{Binding IsBusy}"
+                 EnableNext="True" NextContent="Continue">
+    <StackPanel DockPanel.Dock="Top" Margin="0 0 0 20">
+      <Label Content="Wallet _Name:" Target="walletNameTb" />
+      <TextBox x:Name="walletNameTb" MaxLength="50" Text="{Binding WalletName}" Watermark="(e.g. My Wallet)">
+        <i:Interaction.Behaviors>
+          <behaviors:FocusOnAttachedBehavior />
+        </i:Interaction.Behaviors>
+      </TextBox>
+    </StackPanel>
+  </c:ContentArea>
+</UserControl>
